server/requirements.txt
aliyun-python-sdk-core==2.13.10
amqp==2.5.1
asn1crypto==0.24.0
Babel==2.7.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
billiard==3.6.1.0
celery==4.3.0
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
click==6.7
configparser==5.0.0
cryptography==2.7
DBUtils==1.3
elasticsearch==6.3.1
Flask==1.0.2
Flask-Session==0.3.1
flower==0.9.3
future==0.18.2
get==1.0.3
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
importlib-metadata==0.23
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.4
kombu==4.6.4
lxml==4.3.3
MarkupSafe==1.0
more-itertools==7.2.0
mysql-connector-python==8.0.12
Naked==0.1.31
paramiko==2.6.0
post==1.0.2
protobuf==3.6.1
public==1.0.3
pycparser==2.19
pycryptodome==3.8.0
pyDes==2.0.1
PyNaCl==1.3.0
pytz==2019.2
PyYAML==5.1
query-string==1.0.2
redis==3.3.8
requests==2.19.1
shellescape==3.4.1
six==1.11.0
soupsieve==1.9.1
SpiffWorkflow==0.5.22
tornado==5.1
urllib3==1.23
vine==1.3.0
webencodings==0.5.1
Werkzeug==0.14.1
zipp==0.6.0
